Rename ClientApiGeneratorv2
